+++ packages/sp/src/sharepointqueryable.ts
@@ -132,6 +132,7 @@ export class SharePointQueryable<GetType = any> extends ODataQueryable<SPBatch, 
      */
     protected clone<T extends SharePointQueryable>(factory: SharePointQueryableConstructor<T>, additionalPath?: string, includeBatch = true): T {
         let clone = new factory(this, additionalPath);
+        clone.configure(this._options);
         const target = this.query.get("@target");
         if (target !== null) {
             clone.query.add("@target", target);
@@ -224,25 +225,15 @@ export class SharePointQueryableCollection<GetType = any[]> extends SharePointQu
     }
 
     /**
-     * Orders based on the supplied fields ascending
+     * Orders based on the supplied fields
      *
-     * @param orderby The name of the field to sort on
+     * @param orderby The name of the field on which to sort
      * @param ascending If false DESC is appended, otherwise ASC (default)
      */
     public orderBy(orderBy: string, ascending = true): this {
-        const keys = this._query.getKeys();
-        const query: string[] = [];
-        const asc = ascending ? " asc" : " desc";
-        for (let i = 0; i < keys.length; i++) {
-            if (keys[i] === "$orderby") {
-                query.push(this._query.get("$orderby"));
-                break;
-            }
-        }
-        query.push(`${orderBy}${asc}`);
-
+        const query = this._query.getKeys().filter(k => k === "$orderby").map(k => this._query.get(k));
+        query.push(`${orderBy} ${ascending ? "asc" : "desc"}`);
         this._query.add("$orderby", query.join(","));
-
         return this;
     }